+++ prql-compiler/src/ast/pl/expr/ast.rs
@@ -1,0 +1,315 @@
+use std::collections::HashMap;
+
+use enum_as_inner::EnumAsInner;
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    ast::pl::{Ident, Lineage, Literal, Ty},
+    Span,
+};
+
+use super::TransformCall;
+
+/// Expr is anything that has a value and thus a type.
+/// If it cannot contain nested Exprs, is should be under [ExprKind::Literal].
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Expr {
+    /// Unique identificator of the node. Set exactly once during semantic::resolve.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub id: Option<usize>,
+    #[serde(flatten)]
+    pub kind: ExprKind,
+    #[serde(skip)]
+    pub span: Option<Span>,
+
+    /// For [Ident]s, this is id of node referenced by the ident
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub target_id: Option<usize>,
+
+    /// For [ExprKind::All], these are ids of included nodes
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub target_ids: Vec<usize>,
+
+    /// Type of expression this node represents.
+    /// [None] means that type should be inferred.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ty: Option<Ty>,
+
+    /// Information about where data of this expression will come from.
+    ///
+    /// Currently, this is used to infer relational pipeline frames.
+    /// Must always exists if ty is a relation.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub lineage: Option<Lineage>,
+
+    #[serde(skip)]
+    pub needs_window: bool,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub alias: Option<String>,
+
+    /// When true on [ExprKind::Tuple], this list will be flattened when placed
+    /// in some other list.
+    // TODO: maybe we should have a special ExprKind instead of this flag?
+    #[serde(skip)]
+    pub flatten: bool,
+}
+
+impl Expr {
+    pub fn new(kind: ExprKind) -> Self {
+        Expr {
+            id: None,
+            kind,
+            span: None,
+            target_id: None,
+            target_ids: Vec::new(),
+            ty: None,
+            lineage: None,
+            needs_window: false,
+            alias: None,
+            flatten: false,
+        }
+    }
+
+    pub fn null() -> Expr {
+        Expr::new(ExprKind::Literal(Literal::Null))
+    }
+}
+
+#[derive(Debug, EnumAsInner, PartialEq, Clone, Serialize, Deserialize, strum::AsRefStr)]
+pub enum ExprKind {
+    Ident(Ident),
+    All {
+        within: Ident,
+        except: Vec<Expr>,
+    },
+    Literal(Literal),
+    Pipeline(Pipeline),
+
+    Tuple(Vec<Expr>),
+    Array(Vec<Expr>),
+    Range(Range),
+    Binary(BinaryExpr),
+    Unary(UnaryExpr),
+    FuncCall(FuncCall),
+    Func(Box<Func>),
+    TransformCall(TransformCall),
+    SString(Vec<InterpolateItem>),
+    FString(Vec<InterpolateItem>),
+    Case(Vec<SwitchCase>),
+    RqOperator {
+        name: String,
+        args: Vec<Expr>,
+    },
+
+    Type(Ty),
+
+    /// placeholder for values provided after query is compiled
+    Param(String),
+
+    /// When used instead of function body, the function will be translated to a RQ operator.
+    /// Contains ident of the RQ operator.
+    Internal(String),
+}
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct BinaryExpr {
+    pub left: Box<Expr>,
+    pub op: BinOp,
+    pub right: Box<Expr>,
+}
+
+#[derive(
+    Debug,
+    PartialEq,
+    Eq,
+    Clone,
+    Copy,
+    Hash,
+    Serialize,
+    Deserialize,
+    strum::Display,
+    strum::EnumString,
+)]
+pub enum BinOp {
+    #[strum(to_string = "*")]
+    Mul,
+    #[strum(to_string = "//")]
+    DivInt,
+    #[strum(to_string = "/")]
+    DivFloat,
+    #[strum(to_string = "%")]
+    Mod,
+    #[strum(to_string = "+")]
+    Add,
+    #[strum(to_string = "-")]
+    Sub,
+    #[strum(to_string = "==")]
+    Eq,
+    #[strum(to_string = "!=")]
+    Ne,
+    #[strum(to_string = ">")]
+    Gt,
+    #[strum(to_string = "<")]
+    Lt,
+    #[strum(to_string = ">=")]
+    Gte,
+    #[strum(to_string = "<=")]
+    Lte,
+    #[strum(to_string = "~=")]
+    RegexSearch,
+    #[strum(to_string = "&&")]
+    And,
+    #[strum(to_string = "||")]
+    Or,
+    #[strum(to_string = "??")]
+    Coalesce,
+}
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct UnaryExpr {
+    pub op: UnOp,
+    pub expr: Box<Expr>,
+}
+
+#[derive(
+    Debug,
+    PartialEq,
+    Eq,
+    Clone,
+    Copy,
+    Hash,
+    Serialize,
+    Deserialize,
+    strum::Display,
+    strum::EnumString,
+)]
+pub enum UnOp {
+    #[strum(to_string = "-")]
+    Neg,
+    #[strum(to_string = "+")]
+    Add, // TODO: rename to Pos
+    #[strum(to_string = "!")]
+    Not,
+    #[strum(to_string = "==")]
+    EqSelf,
+}
+
+/// Function call.
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct FuncCall {
+    pub name: Box<Expr>,
+    pub args: Vec<Expr>,
+    #[serde(default, skip_serializing_if = "HashMap::is_empty")]
+    pub named_args: HashMap<String, Expr>,
+}
+
+impl FuncCall {
+    pub fn new_simple(name: Expr, args: Vec<Expr>) -> Self {
+        FuncCall {
+            name: Box::new(name),
+            args,
+            named_args: HashMap::new(),
+        }
+    }
+}
+
+/// Function called with possibly missing positional arguments.
+/// May also contain environment that is needed to evaluate the body.
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct Func {
+    /// Name of the function. Used for user-facing messages only.
+    pub name_hint: Option<Ident>,
+
+    pub return_ty_expr: Option<Expr>,
+
+    /// Type requirement for the function body expression.
+    pub return_ty: Option<Ty>,
+
+    /// Expression containing parameter (and environment) references.
+    pub body: Box<Expr>,
+
+    /// Positional function parameters.
+    pub params: Vec<FuncParam>,
+
+    /// Named function parameters.
+    pub named_params: Vec<FuncParam>,
+
+    /// Arguments that have already been provided.
+    pub args: Vec<Expr>,
+
+    /// Additional variables that the body of the function may need to be
+    /// evaluated.
+    pub env: HashMap<String, Expr>,
+}
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct FuncParam {
+    pub name: String,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ty_expr: Option<Expr>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ty: Option<Ty>,
+
+    pub default_value: Option<Box<Expr>>,
+}
+
+/// A value and a series of functions that are to be applied to that value one after another.
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct Pipeline {
+    pub exprs: Vec<Expr>,
+}
+
+impl From<Vec<Expr>> for Pipeline {
+    fn from(nodes: Vec<Expr>) -> Self {
+        Pipeline { exprs: nodes }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
+pub enum InterpolateItem<T = Expr> {
+    String(String),
+    Expr {
+        expr: Box<T>,
+        format: Option<String>,
+    },
+}
+
+/// Inclusive-inclusive range.
+/// Missing bound means unbounded range.
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
+pub struct Range<T = Box<Expr>> {
+    pub start: Option<T>,
+    pub end: Option<T>,
+}
+
+impl<T> Range<T> {
+    pub const fn unbounded() -> Self {
+        Range {
+            start: None,
+            end: None,
+        }
+    }
+
+    pub fn try_map<U, E, F: Fn(T) -> Result<U, E>>(self, f: F) -> Result<Range<U>, E> {
+        Ok(Range {
+            start: self.start.map(&f).transpose()?,
+            end: self.end.map(f).transpose()?,
+        })
+    }
+
+    pub fn map<U, F: Fn(T) -> U>(self, f: F) -> Range<U> {
+        Range {
+            start: self.start.map(&f),
+            end: self.end.map(f),
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
+pub struct SwitchCase<T = Box<Expr>> {
+    pub condition: T,
+    pub value: T,
+}
